gcf-test/Cloud_run_functions/HTTPTriggers/bq_tble_crf/requirements.txt
function-framework==1.0.0
google-cloud-bigquery==3.5.0
gunicorn==20.1.1
